plugin: test refactoring
extracted utils*_test.go
