+++ pkg/csplugin/broker_test.go
@@ -38,54 +38,6 @@ func setPluginPermTo722(t *testing.T) {
 func setPluginPermTo724(t *testing.T) {
 	setPluginPermTo(t, "724")
 }
-func TestGetPluginNameAndTypeFromPath(t *testing.T) {
-	setUp(t)
-	defer tearDown(t)
-	type args struct {
-		path string
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    string
-		want1   string
-		expectedErr string
-	}{
-		{
-			name: "valid plugin name, single dash",
-			args: args{
-				path: path.Join(testPath, "notification-gitter"),
-			},
-			want:    "notification",
-			want1:   "gitter",
-		},
-		{
-			name: "invalid plugin name",
-			args: args{
-				path: "./tests/gitter",
-			},
-			expectedErr: "plugin name ./tests/gitter is invalid. Name should be like {type-name}",
-		},
-		{
-			name: "valid plugin name, multiple dash",
-			args: args{
-				path: "./tests/notification-instant-slack",
-			},
-			want:    "notification-instant",
-			want1:   "slack",
-		},
-	}
-	for _, tc := range tests {
-		tc := tc
-		t.Run(tc.name, func(t *testing.T) {
-			got, got1, err := getPluginTypeAndSubtypeFromPath(tc.args.path)
-			cstest.RequireErrorContains(t, err, tc.expectedErr)
-
-			assert.Equal(t, tc.want, got)
-			assert.Equal(t, tc.want1, got1)
-		})
-	}
-}
 
 func TestListFilesAtPath(t *testing.T) {
 	setUp(t)
